refactor(web): extract comments-section from task detail
